refactor(core/prodtest): factor out `check_cert_chain`
[no changelog]
